requirements.txt
fastapi
openai
python-dotenv==1.0.1
uvicorn==0.34.0